+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntityComboBoxPanel.java
@@ -25,20 +25,27 @@ import is.codion.swing.common.ui.component.builder.ComponentBuilder;
 import is.codion.swing.common.ui.component.combobox.ComboBoxBuilder;
 import is.codion.swing.common.ui.component.value.AbstractComponentValue;
 import is.codion.swing.common.ui.component.value.ComponentValue;
+import is.codion.swing.common.ui.key.KeyboardShortcuts;
 import is.codion.swing.common.ui.key.TransferFocusOnEnter;
 import is.codion.swing.framework.model.component.EntityComboBoxModel;
 import is.codion.swing.framework.ui.EntityEditPanel;
 
 import javax.swing.Action;
 import javax.swing.JPanel;
+import javax.swing.KeyStroke;
 import java.awt.BorderLayout;
 import java.awt.event.FocusAdapter;
 import java.awt.event.FocusEvent;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.function.Function;
 import java.util.function.Supplier;
 
+import static is.codion.swing.common.ui.key.KeyboardShortcuts.keyStroke;
+import static is.codion.swing.common.ui.key.KeyboardShortcuts.keyboardShortcuts;
 import static is.codion.swing.framework.ui.component.EntityControls.*;
+import static java.awt.event.InputEvent.CTRL_DOWN_MASK;
+import static java.awt.event.KeyEvent.VK_INSERT;
 import static java.util.Objects.requireNonNull;
 
 /**
@@ -46,16 +53,27 @@ import static java.util.Objects.requireNonNull;
  */
 public final class EntityComboBoxPanel extends JPanel {
 
+  public static final KeyboardShortcuts<KeyboardShortcut> KEYBOARD_SHORTCUTS = keyboardShortcuts(KeyboardShortcut.class, new DefaultKeyboardShortcuts());
+
+  /**
+   * The available keyboard shortcuts.
+   */
+  public enum KeyboardShortcut {
+    ADD, EDIT
+  }
+
   private final EntityComboBox comboBox;
 
   private EntityComboBoxPanel(DefaultBuilder builder) {
     comboBox = builder.createComboBox();
     List<Action> actions = new ArrayList<>();
     if (builder.add) {
-      actions.add(createAddControl(comboBox, builder.editPanelSupplier));
+      actions.add(createAddControl(comboBox, builder.editPanelSupplier,
+              builder.keyboardShortcuts.keyStroke(KeyboardShortcut.ADD).get()));
     }
     if (builder.edit) {
-      actions.add(createEditControl(comboBox, builder.editPanelSupplier));
+      actions.add(createEditControl(comboBox, builder.editPanelSupplier,
+              builder.keyboardShortcuts.keyStroke(KeyboardShortcut.EDIT).get()));
     }
     setLayout(new BorderLayout());
     add(createButtonPanel(comboBox, builder.buttonsFocusable, builder.buttonLocation,
@@ -125,6 +143,13 @@ public final class EntityComboBoxPanel extends JPanel {
     Builder buttonLocation(String buttonLocation);
 
     /**
+     * @param keyboardShortcut the keyboard shortcut key
+     * @param keyStroke the keyStroke to assign to the given shortcut key, null resets to the default one
+     * @return this builder instance
+     */
+    Builder keyStroke(KeyboardShortcut keyboardShortcut, KeyStroke keyStroke);
+
+    /**
      * @return a new {@link EntityComboBoxPanel} based on this builder
      */
     EntityComboBoxPanel build();
@@ -144,10 +169,23 @@ public final class EntityComboBoxPanel extends JPanel {
     }
   }
 
+  private static final class DefaultKeyboardShortcuts implements Function<KeyboardShortcut, KeyStroke> {
+
+    @Override
+    public KeyStroke apply(KeyboardShortcut shortcut) {
+      switch (shortcut) {
+        case ADD: return keyStroke(VK_INSERT);
+        case EDIT: return keyStroke(VK_INSERT, CTRL_DOWN_MASK);
+        default: throw new IllegalArgumentException();
+      }
+    }
+  }
+
   private static final class DefaultBuilder extends AbstractComponentBuilder<Entity, EntityComboBoxPanel, Builder> implements Builder {
 
     private final ComboBoxBuilder<Entity, EntityComboBox, ?> entityComboBoxBuilder;
     private final Supplier<EntityEditPanel> editPanelSupplier;
+    private final KeyboardShortcuts<KeyboardShortcut> keyboardShortcuts = KEYBOARD_SHORTCUTS.copy();
 
     private boolean add;
     private boolean edit;
@@ -181,6 +219,12 @@ public final class EntityComboBoxPanel extends JPanel {
     @Override
     public Builder buttonLocation(String buttonLocation) {
       this.buttonLocation = validateButtonLocation(buttonLocation);
+      return this;
+    }
+
+    @Override
+    public Builder keyStroke(KeyboardShortcut keyboardShortcut, KeyStroke keyStroke) {
+      keyboardShortcuts.keyStroke(keyboardShortcut).set(keyStroke);
       return this;
     }